+++ core/src/main/java/org/opensearch/sql/executor/execution/ContinuePaginatedPlan.java
@@ -7,9 +7,9 @@ package org.opensearch.sql.executor.execution;
 
 import org.opensearch.sql.common.response.ResponseListener;
 import org.opensearch.sql.executor.ExecutionEngine;
-import org.opensearch.sql.executor.PaginatedPlanCache;
 import org.opensearch.sql.executor.QueryId;
 import org.opensearch.sql.executor.QueryService;
+import org.opensearch.sql.executor.pagination.PaginatedPlanCache;
 import org.opensearch.sql.planner.physical.PhysicalPlan;
 
 /**
