+++ core/src/main/java/org/opensearch/sql/executor/pagination/CanPaginateVisitor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.sql.executor;
+package org.opensearch.sql.executor.pagination;
 
 import org.opensearch.sql.ast.AbstractNodeVisitor;
 import org.opensearch.sql.ast.Node;
