+++ core/src/main/java/org/opensearch/sql/executor/pagination/Cursor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.sql.opensearch.executor;
+package org.opensearch.sql.executor.pagination;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
