+++ core/src/main/java/org/opensearch/sql/executor/pagination/PaginatedPlanCache.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.sql.executor;
+package org.opensearch.sql.executor.pagination;
 
 import com.google.common.hash.HashCode;
 import java.io.ByteArrayInputStream;
@@ -17,7 +17,6 @@ import lombok.RequiredArgsConstructor;
 import org.opensearch.sql.ast.tree.UnresolvedPlan;
 import org.opensearch.sql.expression.NamedExpression;
 import org.opensearch.sql.expression.serialization.DefaultExpressionSerializer;
-import org.opensearch.sql.opensearch.executor.Cursor;
 import org.opensearch.sql.planner.physical.PaginateOperator;
 import org.opensearch.sql.planner.physical.PhysicalPlan;
 import org.opensearch.sql.planner.physical.ProjectOperator;
