+++ protocol/src/main/java/org/opensearch/sql/protocol/response/QueryResult.java
@@ -16,7 +16,7 @@ import org.opensearch.sql.data.model.ExprValue;
 import org.opensearch.sql.data.model.ExprValueUtils;
 import org.opensearch.sql.executor.ExecutionEngine;
 import org.opensearch.sql.executor.ExecutionEngine.Schema.Column;
-import org.opensearch.sql.opensearch.executor.Cursor;
+import org.opensearch.sql.executor.pagination.Cursor;
 
 /**
  * Query response that encapsulates query results and isolate {@link ExprValue}
